# server/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MediaTracker",
        "version": "0.0.1",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "security": [
        {
            "ApiKey": []
        },
        {
            "ApiHeader": []
        }
    ],
    "paths": {
...
        "/api/user/login": {
            "post": {
                "tags": ["User"],
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "required": ["password", "username"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/text": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "tags": ["User"],
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "confirmPassword": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "confirmPassword",
                                    "password",
                                    "username"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/UserResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/RequestError"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/notification-credentials": {
            "get": {
                "tags": ["User"],
                "operationId": "getNotificationCredentials",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "gotify": {
                                            "type": "object",
                                            "properties": {
                                                "url": {
                                                    "type": "string"
                                                },
                                                "token": {
                                                    "type": "string"
                                                },
                                                "priority": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "priority",
                                                "token",
                                                "url"
                                            ],
                                            "nullable": true
                                        },
                                        "Pushbullet": {
                                            "type": "object",
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": ["token"],
                                            "nullable": true
                                        },
                                        "Pushover": {
                                            "type": "object",
                                            "properties": {
                                                "key": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": ["key"],
                                            "nullable": true
                                        },
                                        "Pushsafer": {
                                            "type": "object",
                                            "properties": {
                                                "key": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": ["key"],
                                            "nullable": true
                                        },
                                        "ntfy": {
                                            "type": "object",
                                            "properties": {
                                                "url": {
                                                    "type": "string"
                                                },
                                                "priority": {
                                                    "type": "string"
                                                },
                                                "topic": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "priority",
                                                "topic",
                                                "url"
                                            ],
                                            "nullable": true
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["User"],
                "operationId": "updateNotificationCredentials",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NotificationPlatformsResponseType"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/text": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/user/settings": {
            "put": {
                "tags": ["User"],
                "operationId": "update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "publicReviews": {
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "sendNotificationWhenStatusChanges": {
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "sendNotificationWhenReleaseDateChanges": {
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "sendNotificationWhenNumberOfSeasonsChanges": {
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "sendNotificationForReleases": {
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "sendNotificationForEpisodesReleases": {
                                        "type": "boolean",
                                        "nullable": true
                                    },
                                    "notificationPlatform": {
                                        "enum": [
                                            "Pushbullet",
                                            "Pushover",
                                            "Pushsafer",
                                            "gotify",
                                            "ntfy",
                                            null
                                        ],
                                        "type": "string",
                                        "nullable": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/text": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/user/password": {
            "put": {
                "tags": ["User"],
                "operationId": "updatePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "currentPassword": {
                                        "type": "string"
                                    },
                                    "newPassword": {
                                        "type": "string"
                                    }
                                },
                                "required": ["currentPassword", "newPassword"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/text": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/user/{userId}": {
            "get": {
                "tags": ["User"],
                "operationId": "getById",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "number"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["id", "name"],
                                    "nullable": true
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKey": {
                "type": "apiKey",
                "in": "query",
                "name": "token"
            },
            "ApiHeader": {
                "type": "apiKey",
                "in": "header",
                "name": "Access-Token"
            }
        },
        "schemas": {
...
            "UserResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "admin": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "publicReviews": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "sendNotificationWhenStatusChanges": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "sendNotificationWhenReleaseDateChanges": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "sendNotificationWhenNumberOfSeasonsChanges": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "sendNotificationForReleases": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "sendNotificationForEpisodesReleases": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "notificationPlatform": {
                        "enum": [
                            "Pushbullet",
                            "Pushover",
                            "Pushsafer",
                            "gotify",
                            "ntfy",
                            null
                        ],
                        "type": "string",
                        "nullable": true
                    }
                },
                "required": ["id", "name"]
            },
            "RequestError": {
                "type": "object",
                "properties": {
                    "errorMessage": {
                        "type": "string"
                    },
                    "MediaTrackerError": {
                        "type": "boolean",
                        "enum": [true]
                    }
                },
                "required": ["MediaTrackerError", "errorMessage"]
            },
            "NotificationPlatformsResponseType": {
                "oneOf": [
                    {
                        "type": "object",
                        "properties": {
                            "platformName": {
                                "type": "string",
                                "enum": ["gotify"]
                            },
                            "credentials": {
                                "type": "object",
                                "properties": {
                                    "url": {
                                        "type": "string"
                                    },
                                    "token": {
                                        "type": "string"
                                    },
                                    "priority": {
                                        "type": "string"
                                    }
                                },
                                "required": ["priority", "token", "url"]
                            }
                        },
                        "required": ["credentials", "platformName"]
                    },
                    {
                        "type": "object",
                        "properties": {
                            "platformName": {
                                "type": "string",
                                "enum": ["Pushbullet"]
                            },
                            "credentials": {
                                "type": "object",
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    }
                                },
                                "required": ["token"]
                            }
                        },
                        "required": ["credentials", "platformName"]
                    },
                    {
                        "type": "object",
                        "properties": {
                            "platformName": {
                                "type": "string",
                                "enum": ["Pushover"]
                            },
                            "credentials": {
                                "type": "object",
                                "properties": {
                                    "key": {
                                        "type": "string"
                                    }
                                },
                                "required": ["key"]
                            }
                        },
                        "required": ["credentials", "platformName"]
                    },
                    {
                        "type": "object",
                        "properties": {
                            "platformName": {
                                "type": "string",
                                "enum": ["Pushsafer"]
                            },
                            "credentials": {
                                "type": "object",
                                "properties": {
                                    "key": {
                                        "type": "string"
                                    }
                                },
                                "required": ["key"]
                            }
                        },
                        "required": ["credentials", "platformName"]
                    },
                    {
                        "type": "object",
                        "properties": {
                            "platformName": {
                                "type": "string",
                                "enum": ["ntfy"]
                            },
                            "credentials": {
                                "type": "object",
                                "properties": {
                                    "url": {
                                        "type": "string"
                                    },
                                    "priority": {
                                        "type": "string"
                                    },
                                    "topic": {
                                        "type": "string"
                                    }
                                },
                                "required": ["priority", "topic", "url"]
                            }
                        },
                        "required": ["credentials", "platformName"]
                    }
                ]
            }
        }
    }
}